React-Native 0.76 upgrade (#199)
* feat: updgrade react native to 0.76

* fix: re-rendering issues on session scoring

* fix: versioning and react-native-paper button issues

* chore: lock update

* build: upgrade rn version ios changes (#196)

* fix: (6041455348) logged in user object empty

* fix: (6043198873) react native screens additional steps

* fix: issue on signup invalid email address for longer email affix

---------

+++ src/auth/AuthProvider/AuthProvider.tsx
@@ -89,14 +89,12 @@ const AuthProvider = ({ children, storage }: Props) => {
 	};
 
 	const signOut = async () => {
-		const { id } = loggedInUser as LoginResponseSchemaType;
-
 		// remove push token from server
 		try {
-			if (pushToken) {
+			if (pushToken && loggedInUser?.id) {
 				await deletePushToken(
 					pushToken,
-					id,
+					loggedInUser.id,
 					Func.getEnv(
 						storage.getString('apiUrl') || Constant.API_URL,
 					),
